fix(signal): four more the second adversarial pass found
Each reproduced with a probe before the fix; four new regression tests, each
born failing.

  * LpgT could BOOST, which makes its "can only attenuate" contract — and the
    worst_case_gain of 1.0 the Forge registry declares for it, which a path-gain
    lint sums over a chain — untrue. Above sample_rate / 4 the trapezoidal
    one-pole's step response exceeds one: near-Nyquist content pumps the
    integrator past the input bound and the next low-frequency sample reads it
    back out as a peak above the input's, up to ~1.5x at 18 kHz / 44.1 kHz. The
    commanded cutoff is now capped at sample_rate / 4, below which every state
    update is a convex combination and peak-out <= peak-in holds
    unconditionally. This also fixes the gate at low sample rates, where the
    default 12 kHz ceiling sits above Nyquist outright.
  * SlewLimiterT's accumulator was SampleType. In linear mode a long ramp at a
    high sample rate steps by less than half an ulp of a float near 1, so the
    accumulator absorbs the step and stalls short of the target forever — the
    "a full 0 -> 1 step takes exactly rise_ms" contract, broken precisely where
    the times are longest. The accumulator is now double regardless of
    SampleType; the interface is unchanged.
  * OuWalkT integrated its drift with the naive Euler-Maruyama `theta * dt`,
    which matches the exact factor only while that product is small. Past 1 it
    overshoots the mean, and past 2 it turns the walk into a clamp-to-clamp
    oscillator — reachable with a strong pull at a low update rate, which is
    exactly what DriftT runs (its control rate is the sample rate over 32, so
    its dt is 32x larger than a per-sample walk's). It now uses the exact
    per-step decay `1 - e^(-theta*dt)`, a fraction in (0, 1] for every theta,
    so the pull is unconditionally stable.
  * LfoT::cycles_ was a 32-bit int. A free-running LFO near the rate ceiling
    completes ~86400 cycles a second at 192 kHz and overflows it within hours —
    a long-running instance, which is what "free-running" means.

+++ core/signal/include/pulp/signal/lfo.hpp
@@ -236,7 +236,7 @@ public:
 
     Stage stage() const { return stage_; }
     bool finished() const { return stage_ == Stage::done; }
-    int cycles_completed() const { return cycles_; }
+    long long cycles_completed() const { return cycles_; }
     double phase() const { return phase_; }
 
     // ── output ───────────────────────────────────────────────────────────────
@@ -420,7 +420,7 @@ private:
         if (phase_ >= 1.0) {
             const double whole = std::floor(phase_);
             phase_ -= whole;
-            cycles_ += static_cast<int>(whole);
+            cycles_ += static_cast<long long>(whole);
             on_cycles_completed_();
         }
     }
@@ -531,7 +531,9 @@ private:
     std::uint32_t seed_ = 0x12345678u;
     int segments_ = 1;
     int repeat_count_ = 0;
-    int cycles_ = 0;
+    // 64-bit: a free-running LFO at the rate ceiling wraps ~86400 cycles per
+    // second at 192 kHz, which overflows a 32-bit counter within hours.
+    long long cycles_ = 0;
 
     Wave wave_ = Wave::sine;
     Mode mode_ = Mode::free;
